adding new sub directives to manage stages and status
