+++ app/post/post-module.js
@@ -11,6 +11,8 @@ angular.module('ushahidi.posts', [])
 .directive('postEditor', require('./directives/post-editor-directive.js'))
 .directive('postChooseForm', require('./directives/post-choose-form-directive.js'))
 .directive('postValueEdit', require('./directives/post-value-edit-directive.js'))
+.directive('postStatus', require('./directives/post-status-directive.js'))
+.directive('postStages', require('./directives/post-stages-directive.js'))
 // Views
 .directive('postView', require('./directives/views/post-view-directive.js'))
 .directive('postViewMap', require('./directives/views/post-view-map-directive.js'))
