+++ server/www/templates/posts/post-status.html
@@ -1,0 +1,5 @@
+<div ng-show="allowedChangeStatus()" class="form-field select">
+    <role-selector
+        toggle-role-func="publishPostTo(updatedPost)"
+        post="post" ></role-selector>
+</div>
